requirements.txt
PyYAML>=6.0
isort==6.0.1; python_full_version >= '3.9.0'
